Added ability to mount report volume
This update will mount and unmount report_vol, as well as adding or
removing the mount point depending on the service status, in order to
place the source of truth inside puppet rather than having it somewhere
else. If you don't want the module to handle mounting, you can leave the
fstab_entry as undef.